api/requirements.txt
fastapi
uvicorn
openai>=1.0.0
pydantic
python-dotenv
pdfplumber
scikit-learn
numpy
